feat(EVM-1167): cleanup storage types module (#543)
## What ❔
- Removed unused storage-type markers/bitflags.
- Split `storage_types` into focused modules.
- Moved `MAX_EVENT_TOPICS` from storage types into system constants and
updated imports.

## Why ❔
- Reduces dead code and improves module structure.
- Places constants with the correct domain ownership for easier
maintenance.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist
- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted.
