+++ basic_bootloader/src/bootloader/transaction_flow/ethereum/mod.rs
@@ -18,12 +18,12 @@ use zk_ee::execution_environment_type::ExecutionEnvironmentType;
 use zk_ee::interface_error;
 use zk_ee::internal_error;
 use zk_ee::out_of_native_resources;
-use zk_ee::storage_types::MAX_EVENT_TOPICS;
 use zk_ee::system::errors::root_cause::GetRootCause;
 use zk_ee::system::errors::root_cause::RootCause;
 use zk_ee::system::errors::runtime::RuntimeError;
 use zk_ee::system::errors::subsystem::SubsystemError;
 use zk_ee::system::logger::Logger;
+use zk_ee::system::MAX_EVENT_TOPICS;
 use zk_ee::system::*;
 use zk_ee::system_log;
 use zk_ee::types_config::EthereumIOTypesConfig;
