+++ evm_interpreter/src/instructions/host.rs
@@ -5,9 +5,9 @@ use gas_constants::{CALL_STIPEND, INITCODE_WORD_COST, SHA3WORD};
 
 use native_resource_constants::*;
 use zk_ee::common_structs::system_hooks::HooksStorage;
-use zk_ee::storage_types::MAX_EVENT_TOPICS;
 use zk_ee::system::tracer::evm_tracer::EvmTracer;
 use zk_ee::system::tracer::Tracer;
+use zk_ee::system::MAX_EVENT_TOPICS;
 use zk_ee::{system::*, wrap_error};
 
 use super::*;
