+++ zk_ee/src/storage_types/storage_address.rs
@@ -1,0 +1,31 @@
+use crate::oracle::usize_serialization::{UsizeDeserializable, UsizeSerializable};
+use crate::utils::exact_size_chain::ExactSizeChain;
+use crate::{system::errors::internal::InternalError, types_config::SystemIOTypesConfig};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+pub struct StorageAddress<IOTypes: SystemIOTypesConfig> {
+    pub address: IOTypes::Address,
+    pub key: IOTypes::StorageKey,
+}
+
+impl<IOTypes: SystemIOTypesConfig> UsizeSerializable for StorageAddress<IOTypes> {
+    const USIZE_LEN: usize = <IOTypes::Address as UsizeSerializable>::USIZE_LEN
+        + <IOTypes::StorageKey as UsizeSerializable>::USIZE_LEN;
+
+    fn iter(&self) -> impl ExactSizeIterator<Item = usize> {
+        ExactSizeChain::new(
+            UsizeSerializable::iter(&self.address),
+            UsizeSerializable::iter(&self.key),
+        )
+    }
+}
+
+impl<IOTypes: SystemIOTypesConfig> UsizeDeserializable for StorageAddress<IOTypes> {
+    const USIZE_LEN: usize = <Self as UsizeSerializable>::USIZE_LEN;
+
+    fn from_iter(src: &mut impl ExactSizeIterator<Item = usize>) -> Result<Self, InternalError> {
+        let address = UsizeDeserializable::from_iter(src)?;
+        let key = UsizeDeserializable::from_iter(src)?;
+        Ok(Self { address, key })
+    }
+}
